Add files via upload
By analyzing Tesla's revenue from 2008 to 2022, we see that with the Mean model there is no change despite the number of periods. As we move to the Naive model however, we start to see a slight increase over each period. This is nowhere near the significant increase we see year over year from the actual data.